% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{bracket_hours}
\alias{bracket_hours}
\title{Bracket Hours.}
\usage{
bracket_hours(x, bracket_idx = 1:ncol(x), bracket_mult = NULL, append = TRUE)
}
\arguments{
\item{x}{A data.frame (also) containing the subjective and retrospectively
provided practice times of each participant.}

\item{bracket_idx}{Optional numerical vector of size \code{ncol(x)} indexing
those columns of \code{x} that contain mean daily hours information. Not
necessary if \code{x} only has columns with daily age bracket means.}

\item{bracket_mult}{A numerical vector giving the durations of each age
bracket.}

\item{append}{A logic value determining whether the function is to return the
yearly hours appended to \code{x} (TRUE; default), or just the yearly
hours.}
}
\value{
If \code{append = TRUE} a data.frame of size \code{nrow(x) x (ncol(x)
  + length(bracket_idx))}. If \code{append = FALSE} a data.frame of size
\code{nrow(x) x length(bracket_idx)}.
}
\description{
Calculates the hours subjects have been practice during age brackets.
}
\examples{
d_f <- data.frame(ID = c(1, 2), `10-12` = c(1.5, 1.0), `13-14` = c(3, 0.75), check.names = FALSE)
index <- c(2, 3)
bracket_hours(d_f, index, c(3, 2))
}
