% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{bracket_multipliers}
\alias{bracket_multipliers}
\title{Age Bracket Multiplier.}
\usage{
bracket_multipliers(df, bracket_idx = 1:ncol(df), sep = "-")
}
\arguments{
\item{df}{A data.frame (also) containing the subjective and retrospectively
provided practice times of each participant.}

\item{bracket_idx}{A numerical vector indexing the bracket columns. If not
provided, the entire data.frame is assumed to only contain age bracket
columns.}

\item{sep}{A character string used to separate the beginning and ending ages
in age bracket column names.}
}
\value{
A numerical vector the same length as \code{nrow(df)}, providing the
indices of the right-most non-NA entry in each row of \code{df}.
}
\description{
Age Bracket Multiplier.
}
\examples{
d_f <- data.frame(`10-12` = c(1.5, 1.0, 0.75), `13-14` = c(3, 0.75, 1.0), check.names = FALSE)
bracket_multipliers(d_f)
}
