% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_firstentry_}
\alias{col_firstentry_}
\title{First Column With (Non-NA) Content.}
\usage{
col_firstentry_(x, cols = 1:ncol(x))
}
\arguments{
\item{x}{A data.frame.}

\item{cols}{Either a numerical vector indexing the age bracket columns, or a
character vector with age bracket column names.}
}
\value{
A vector with age bracket column indices.
}
\description{
Returns the column index of the left-most non-NA entry of each row in a
data.frame.
}
\examples{
d_f <- data.frame(a = c("a", "b", NA), b = c(1, NA, 3))
col_firstentry(d_f)
}
