% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{col_names_to_indices}
\alias{col_names_to_indices}
\title{Convert Column Names to Column Indices.}
\usage{
col_names_to_indices(x, cols)
}
\arguments{
\item{x}{A data.frame with practice hours for each age bracket, one  line per
subject.}

\item{cols}{Either \code{NULL} (the default), using all columns of x as age
brackets; or a numerical vector indexing the age bracket columns to use; or
a regular expression matching the names of the age bracket columns.}
}
\value{
A vector with indices.
}
\description{
Convert Column Names to Column Indices.
}
