% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{days_lived}
\alias{days_lived}
\title{Days Lived Since Date Of Birth.}
\usage{
days_lived(dob, dte = Sys.Date())
}
\arguments{
\item{dob}{Date of birth.}

\item{dte}{If not provided, the value returned by \code{Sys.Date()}. The date
between which and \code{dob} the age shall be calculated.}
}
\value{
The age as \code{lubridate} period.
}
\description{
Calculate age in days from date of birth and, e.g. today's date, to day
precision.
}
\examples{
days_lived(lubridate::ymd("1995-01-28"), lubridate::ymd("2021-01-27")) # "25y 11m 30d 0H 0M 0S"
days_lived(as.Date("2020-02-28"), as.Date("2021-02-27")) # "11m 30d 0H 0M 0S"
}
\author{
Garret Grolemund, https://stackoverflow.com/users/436363/garrett
https://stackoverflow.com/questions/3611314/calculate-ages-in-r
}
