% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{exact_duration}
\alias{exact_duration}
\title{Exact, Leap Year-aware Duration.}
\usage{
exact_duration(
  x,
  cols = 1:ncol(x),
  align = NULL,
  dos = NULL,
  dob = NULL,
  pract_start = NULL
)
}
\arguments{
\item{x}{A data.frame (also) containing subjective and retrospectively
provided practice times, one row per participant.}

\item{cols}{A numerical vector of size \code{ncol(x)} indexing those columns
of \code{x} that contain mean daily hours information; or a character
string with the names of those columns containing age bracket averages. If
omitted, all columns are assumed to contained practice time data.}

\item{align}{Either a character string determining whether to align the data
set with each participants' first (\code{align = "first_entry"}) or last
entry (\code{align = "last_entry"}), or a numeric vector indexing each
participant's column with which to align the data set.}

\item{dos}{Optional (character string) column name with dates of filling in
the questionnaire (date of study).}

\item{dob}{Optional (character string) column name with dates of birth.}

\item{dop}{Optional (character string) column name with dates of
practice commencement (date of practice).}
}
\value{
A \code{data.frame} with exact practice times.
}
\description{
Exact, Leap Year-aware Duration.
}
