% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{extend_bracket}
\alias{extend_bracket}
\title{Extend Bracket.}
\usage{
extend_bracket(x, colnames, dur, grp)
}
\arguments{
\item{x}{A long-format data.frame.}

\item{cols}{A vector with age bracket column names of the wide-format
data.frame.}

\item{duration}{A numerical vector with age bracket durations.}
}
\value{

}
\description{
Replace Age Bracket By Repeating Data n-Times.
}
