% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{names_to_durations}
\alias{names_to_durations}
\title{Convert Column Names to Age Bracket Durations.}
\usage{
names_to_durations(col_names = NULL, sep = NULL)
}
\arguments{
\item{col_names}{A character vector of age brackets.}

\item{sep}{A character used to separate the lower and upper bound of the
bracket ranges.}
}
\value{
A numerical vector with age bracket durations (usually in years).
}
\description{
Convert Column Names to Age Bracket Durations.
}
\examples{
d_f <- data.frame(`10-12` = c(1.5, 1.0, 0.75), `13-14` = c(3, 0.75, 1.0), check.names = FALSE)
names_to_durations(names(d_f), sep = "-")
}
