% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{practicetime}
\alias{practicetime}
\title{Practicing Times of Musicians.}
\format{
A data frame with 68 rows and 10 variables:
\describe{
\item{ID}{subject ID}
\item{DOB}{Date of birth, in YYYY-MM-DD format}
\item{Date}Date of lab visit}
\item{0-5}{Mean daily practicing time in the age bracket 0 - 5}
\item{6-8}{Mean daily practicing time in the age bracket 6 - 8}
\item{9-12}{Mean daily practicing time in the age bracket 9 - 12}
\item{13-16}{Mean daily practicing time in the age bracket 13 - 16}
\item{17-20}{Mean daily practicing time in the age bracket 17 - 20}
\item{21-25}{Mean daily practicing time in the age bracket 21 - 25}
\item{26-30}{Mean daily practicing time in the age bracket 26 - 30}
}
\usage{
practicetime
}
\description{
Students at a conservatory were asked to retrospectively estimate the average
daily time spent at the (main) instrument in seven age brackets.
}
\keyword{datasets}
