% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temporal_align_}
\alias{temporal_align_}
\title{Pad \code{data.frame} Rows with \code{NA} Values.}
\usage{
temporal_align_(x, cols = 1:ncol(x), padd = numeric(nrow(x)), side = NULL)
}
\arguments{
\item{x}{A data.frame (also) containing the subjective and retrospectively
provided practice times, one row per participant.}

\item{cols}{A numerical vector of size \code{ncol(x)} indexing those columns
of \code{x} that contain mean daily hours information; or a character
string with the names of those columns containing age bracket averages. If
omitted, all columns are assumed to contained practice time data.}

\item{padd}{A numerical vector of length \code{nrow(x)} with the number of
\code{NA}s to pad each ro with.}

\item{side}{Either a character string determining whether to align the data
set with the participants' right-most ("last_entry") or left-most entry
("first_entry"); or if \code{side} is numeric it is meant to be a column
index, on which the data set gets aligned with.}
}
\value{
A \code{data.frame} padded with \code{NA}s on the left and/or on the
right side.
}
\description{
Given a \code{data.frame} with rows that are (at least partially) left-
and/or right-padded with \code{NA} values (i.e. the number of non-\code{NA}
entries differs between rows), and a criterion which returns \code{TRUE} for
each row, this function returns \code{x} left- and/or right-padded with
\code{NA} values.
}
\examples{
d_f <- data.frame(`10-12` = c(NA, 1.5, 1.0, NA),
                  `13-14` = c(1.5, 3, 0.75, 1),
                  `15-16` = c(2.5, 4, 1.5, NA)
                  check.names = FALSE)
temporal_align(d_f, )
}
