% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{names_to_durations}
\alias{names_to_durations}
\title{Convert Column Names to Age Bracket Durations.}
\usage{
names_to_durations(col_names, sep = "-")
}
\arguments{
\item{col_names}{A character vector with the names of those columns of a
data.frame containing age bracket averages.}

\item{sep}{A character used to separate the lower and the upper bound of the
bracket ranges.}
}
\value{
A numerical vector with age bracket durations (usually in years).
}
\description{
Convert Column Names to Age Bracket Durations.
}
\examples{
var_names <- c("10-12", "13-14")
names_to_durations(var_names, sep = "-")
}
