% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{names_to_durations}
\alias{names_to_durations}
\title{Convert Column Names to Age Bracket Durations.}
\usage{
names_to_durations(x, cols = 1:ncol(x), sep = "-")
}
\arguments{
\item{sep}{A character used to separate the lower and upper bound of the
bracket ranges.}

\item{col_names}{A character vector of age brackets.}
}
\value{
A numerical vector with age bracket durations (usually in years).
}
\description{
Convert Column Names to Age Bracket Durations.
}
\examples{
d_f <- data.frame(Names = c("Carl", "Blue", "Perkins"),
                  `10-12` = c(1.5, 1.0, 0.75),
                  `13-14` = c(3, 0.75, 1.0),
                  check.names = FALSE)
names_to_durations(d_f, "^\\\\d{2}-\\\\d{2}", sep = "-")
names_to_durations(d_f, names(d_f)[2:3])
names_to_durations(d_f, 2:3)
}
