% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{plot_it}
\alias{plot_it}
\title{Plot Deliberate Practice Hours Over Age Brackets.}
\usage{
plot_it(x, cols = NULL, ID = "ID", Group = NULL, legend = FALSE)
}
\arguments{
\item{x}{A data.frame with practice hours for each age bracket, one  line per
subject. See examples.}

\item{cols}{A numerical vector indexing the age bracket columns to use, or a
regular expression matching the names of the age bracket columns.}

\item{ID}{A vector of size \code{nrow(x)} with unique subject IDs.}

\item{Group}{A grouping vector of size \code{nrow(x)}.}

\item{legend}{Logical. When \code{FALSE} (default) or \code{"none"}, no legend
is shown. All possible values for \code{ggplot2::theme(legend.position)}
are allowed.}
}
\description{
Plot Deliberate Practice Hours Over Age Brackets.
}
\examples{

}
