% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praktikr.R
\name{plot_it}
\alias{plot_it}
\title{Plot Deliberate Practice Hours Over Age Brackets.}
\usage{
plot_it(x, cols = 1:ncol(x), ID = "ID", Group = NULL, legend = FALSE)
}
\arguments{
\item{x}{A data.frame with practice hours for each age bracket, one line per
subject (wide format).}

\item{cols}{Either a numerical vector indexing the age bracket columns to
use, or a regular expression matching the names of the age bracket columns,
or a character vector with the bracket column names. If not provided, all
columns of \code{x} are treated as containing age bracket information.}

\item{ID}{A vector of size \code{nrow(x)} with unique subject IDs.}

\item{Group}{A grouping vector of size \code{nrow(x)}.}

\item{legend}{Logical. When \code{FALSE} (default) or \code{"none"}, no
legend is shown. All possible values for
\code{ggplot2::theme(legend.position)} are allowed.}
}
\description{
Plot Deliberate Practice Hours Over Age Brackets.
}
\examples{
d_f <- data.frame(ID = LETTERS[1:3],
                  `10-12` = c(1.5, 1.0, 0.75),
                  `13-14` = c(3, 0.75, 1.0),
                  check.names = FALSE)
## Equivalent calls:
plot_it(d_f, "\\\\d{2}-\\\\d{2}")
plot_it(d_f, c("10-12", "13-14"))
plot_it(d_f, c(2, 3))
plot_it(d_f[ , 2:3])
}
